perf: add ChESS-kernel + Radon-SAT microbenches with measured baseline (PERF-01/02)
Isolate the two heaviest hot paths and record baselines in the perf design doc.

- benches/chess_response.rs: time chess_response_u8 at 256²/512²/1024².
  Scalar ~155 Mpix/s vs SIMD ~621 Mpix/s @1024² (3.7–4.0×).
- benches/radon_response.rs: add 1024² dim; i64 vs u32 SAT.
  u32 win +9–16% at upsample=1, collapses at upsample=2.
- benches/common/: shared synth_chessboard, de-duplicating the Radon
  bench's copy (advances SOLID-01).
- docs/design/perf-profiling.md: record measured baseline + the two
  evidence-backed PERF-10 levers (vectorize per-lane μₗ + write-back;
  upsample=2 angular sampling).
- BACKLOG: PERF-01/02 done; PERF-10 refined; SOLID-01 in-progress.

Verified: fmt clean; clippy -D warnings green in default, simd, and
radon-sat-u32 configs; benches compile and run.
